WEB-60: Removing unused plugin icon download functionality.
This commit removes functionality that appears to serve the icon of a plugin, when a specificly crafted request was made.

I've grepped through 1.9G of access log files, using the command below, and did not find any relevant request.

    grep "plugin=" * | grep "img"

This leads me to believe that we can remove this functionality, and reduce the complexity of the code.