fixes #208: plugin version identifier comparision
As the website code has to deal with plugin version identifiers that sometimes predate the stricter Maven syntax, it is using a simple alphabetic comparator. This causes problems when comparing numbers that do not have the same amount of digits (eg 10 < 9).

This commit works around this issue, by zero-padding all values to the same lenght, before applying the alphabetic comparator. This should not affect ordering of the free-format version identifiers to much, while it will ensure that numeric ordering is improved: 10 > 09.