Restoring 'blog' content on project pages
This restores the latest blog items being displayed on each project page. Blogs are taken based on the tags applies to the "ignite realtime blogs" category of Discourse.

As Discourse's RSS feed does not go back very far, and doesn't seem to offer an RSS feed for the combination of 'category' and 'tag', the amount of history that is displayed is limited.