Swap 'about' and 'home' page
The thought behind most of these changes in this commit is to give first-time visitors that are unfamiliar with our projects a better idea what they're looking at.

To do that, the 'about' page has been moved to the home/landing page, the first thing that people see when finding the page. As it describes what we do, I thought that to be fitting with the motivation described above. The text on this page has been modified, to better describe what type of problem-solving sulutions we offer. This can use (a lot more) love - I'm not a great content writer.

What used to be the home/landing page, has now been moved to a category named 'news' (as it basically was just a feed of news articles). It takes the place in the menu that used to be occupied by the 'about' page.

The sidebars on both pages have been modified, primarily to ensure that 'news' and 'in the community' references remain visible on the home/landing page.

Other changes:
- the sidebar on each 'support' page is now made uniform (and contain a few more items)
- re-worked the re-usablility of some sidebar fragments