Restore 'whitepapers' section, add one whitepaper
In a new 'whitepapers' section (now part of the 'articles'), this commit restores an old but updated whitepaper from Jive, describing XMPP.

The text on on the main page that describes XMPP now links to this whitepaper.

Sidebars have been added to link to the whitepaper(s).

The 'support' landing page now lists the whitepapers.