fixes #276: Less bad rendering on Mobile
This improves the rendering of the website on mobile. It does so mainly by replacing the implementation of 'jumbotrons' with a CSS-grid based thing.

The resulting pages are far from perfect on mobile (the last few elements of the menu now disappear, and the sidebar is weirdly aligned to the right).

I could not detect notable differences on non-mobile.