WEB-82: Restore listing of chatroom JID
Under WEB-82, the text explaining how to join the chat room with an XMPP client was removed, to conserve screem real-estate.

It is desirable to have at least the address be listed. As XMPP now supports URIs, we can easily provide a very short sentence with a link in it, that is usable by people that have installed an XMPP client. In this commit, that's placed on the same line as the 'powered by Converse' line, to minimize the space that this extra text takes up.